Add CodeFileExtractorAndMaintainer class responsible for extraction and linking when cloning the repo and for maintaining links on git fetch
